@seanxcwang 2025-02-12
test on H20（peak bandwidth 4TB/s）
```
Config: batch_size=768, seq_len=1024, num_heads=16
Memory bandwidth: 668.77 GB/s
Config: batch_size=768, seq_len=1024, num_heads=32
Memory bandwidth: 685.71 GB/s
Config: batch_size=768, seq_len=1024, num_heads=64
Memory bandwidth: 718.93 GB/s
Config: batch_size=768, seq_len=2048, num_heads=16
Memory bandwidth: 669.80 GB/s
Config: batch_size=768, seq_len=2048, num_heads=32
Memory bandwidth: 678.41 GB/s
Config: batch_size=768, seq_len=2048, num_heads=64
Memory bandwidth: 695.46 GB/s
```
Is there any performance improvement plan for H20, any suggestion will be helpful