@yzh119 2025-02-12
@seanxcwang I'm working on the FA3 version of this kernel, which should also benefit H20, please stay tuned :)