@flyingcow8 2025-04-17
> For large num_local_heads such as 128 (if no TP and no MTP), we create a cluster of size 2, and the upper half (64) and lower half (64) are dispatched to two SMs in a cluster, and we can use software-managed multicasting (for large page size), we leave it to later PR.

@yzh119 what does software-managed multicasting mean and has the PR  been submitted?